refactor: 폴더 구조 변경 (#6)
메인 페이지에서 사용되는 폴더 그룹화를 위한 private folder (_home) 추가
header 역할 컴포넌트, 관련 hook 파일을 route 폴더 내부로 이동
react-query provider 파일을 provider 폴더 내부로 이동
