+++ app/(route)/_home/components/NavigationBar.tsx
@@ -6,7 +6,7 @@ import { useState } from 'react'
 import { faCaretDown } from '@fortawesome/free-solid-svg-icons/faCaretDown'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
-import NavItem from '@/app/components/NavItem'
+import NavItem from '@/app/(route)/_home/components/NavItem'
 
 export default function NavigationBar() {
   const navList = [
