+++ app/(route)/_home/components/ThemeSwitcher.tsx
@@ -5,7 +5,7 @@ import { faMoon } from '@fortawesome/free-solid-svg-icons/faMoon'
 import { faSun } from '@fortawesome/free-solid-svg-icons/faSun'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
-import useLocalStorage from '@/app/hooks/useLocalStorage'
+import useLocalStorage from '@/app/(route)/_home/hooks/useLocalStorage'
 
 export default function ThemeSwitcher() {
   const THEME_NAME = 'stock-saga-theme'
